Create backlog_df.csv
Above code is to create new .csv file. If file already exists, above shouldn't be used.  Trade should start with empty .csv file. Trade information will be added while the bot makes trades. Purpose of this file, is to check if trades as profitable or if trade criteria (such as moving averages of trade intervales) should be adjusted to prevent losses or maximize profit.